+++ TagsCloudContainer/TagsCloudContainer/IFileReader.cs
@@ -1,0 +1,6 @@
+namespace TagsCloudContainer;
+
+public interface IFileReader
+{
+    internal List<string> ReadFile(string path);
+}
