+++ TagsCloudContainer/TagsCloudContainer/TextRectangleContainerProcessor.cs
@@ -5,25 +5,25 @@ namespace TagsCloudContainer;
 
 public class TextRectangleContainerProcessor
 {
-    private readonly IFileHandler _fileHandler;
+    private readonly IFileReader _fileReader;
     private readonly ILayouter _layouter;
     private readonly RectangleSizeCalculator _rectangleSizeCalculator;
-    private readonly List<string> _banWords;
+    private readonly BoringWordsProcessor _boringWordsProcessor;
     
-    public TextRectangleContainerProcessor(IFileHandler fileHandler, ILayouter layouter,
-        RectangleSizeCalculator calculator, List<string> banWords)
+    public TextRectangleContainerProcessor(IFileReader fileReader, ILayouter layouter,
+        RectangleSizeCalculator calculator, BoringWordsProcessor boringWordsProcessor)
     {
-        this._fileHandler = fileHandler;
+        this._fileReader = fileReader;
         this._layouter = layouter;
         this._rectangleSizeCalculator = calculator;
-        this._banWords = banWords;
+        this._boringWordsProcessor = boringWordsProcessor;
     }
     
     public IEnumerable<TextRectangleContainer?> ProcessFile(string path, IWordMeasurer wordMeasurer)
     {
-        var redData = _fileHandler.HandleFile(path, _banWords);
+        var readData = _boringWordsProcessor.WordsToLowerAndRemoveBoringWords(_fileReader.ReadFile(path));
 
-        foreach (var wordInfo in redData)
+        foreach (var wordInfo in readData)
         {
             var rectSize = _rectangleSizeCalculator.CalculateNextRectangleSize(wordInfo, wordMeasurer, out var scale);
             if (rectSize == Size.Empty)
